Module aws_sdk_apigatewaymanagement::error

Modules

forbidden_exception

See ForbiddenException

gone_exception

See GoneException

limit_exceeded_exception

See LimitExceededException

payload_too_large_exception

See PayloadTooLargeException

Structs

DeleteConnectionError

ForbiddenException

The caller is not authorized to invoke this operation.

GetConnectionError

GoneException

The connection with the provided id no longer exists.

LimitExceededException

The client is sending more than the allowed number of requests per unit of time or the WebSocket client side buffer is full.

PayloadTooLargeException

The data has exceeded the maximum size allowed.

PostToConnectionError

Enums

DeleteConnectionErrorKind

GetConnectionErrorKind

PostToConnectionErrorKind